Move AuthenticationException to ignite-api module
